// frontend/src/backend.d.ts
export interface Location {
    country: string;
    city: string;
    state: string;
}
// ...
export interface ContactInfo {
    email: string;
    website: string;
    address: string;
    phone: string;
}
// ...
export interface AmbulanceRequest {
    id: string;
    status: RequestStatus;
    requester: Principal;
    contact: ContactInfo;
    name: string;
    location: Location;
}
// ...
export enum RequestStatus {
    pending = "pending",
    approved = "approved",
    rejected = "rejected"
}